moved userPlaylist to playlistHome and deleted un-needed file
